Just fix some minor build warnings
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@4737 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/MapGuideCommon/Services/SiteManager.cpp
@@ -98,7 +98,7 @@ ACE_THR_FUNC_RETURN CheckServers(void* param)
         }
 
         // Sleep the thread as thread resume/suspend is not supported on all platforms
-        for(size_t i=0;i<sleepTime;i++)
+        for(INT32 i=0;i<sleepTime;i++)
         {
             ACE_OS::sleep(1);
